requirements.txt
# requirements.txt
torch
transformers>=4.40.0
huggingface_hub
Pillow
streamlit
accelerate

# adding the requirements for FastAPI
fastapi
uvicorn[standard]
httpx
